refactor(arch): extract AppCoordinator to break 4 cross-feature import violations (#1956)
Creates routing/app_coordinator.dart as a shared navigation coordinator
so features can navigate to home/events/tickets/account-deletion without
importing each other's coordinators directly.

Violations fixed:
- my_tickets → home: my_tickets_page now imports AppCoordinator
- home → ticket: HomeCoordinator no longer imports TicketEventMeta;
  pushTicketQR moved to AppCoordinator (routing layer)
- settings → consent: consent_detail_sheet moved to common/widgets;
  consent feature re-exports for backward compatibility
- settings → account_deletion: privacy_page calls
  AppCoordinator.startAccountDeletion() instead of reading
  accountDeletionCoordinatorProvider directly
- account_deletion → home: deletion_complete_page now imports
  AppCoordinator

Constraint: routing/ is non-feature infrastructure; features importing
  from routing/ is permitted by feature-first architecture
Rejected: move TicketEventMeta to minglit_kit | larger scope, pure
  routing-layer import is sufficient to fix the violation
Confidence: high
Scope-risk: narrow
