+++ apps/app_user/lib/src/features/home/logic/home_coordinator.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:app_user/src/features/ticket/ui/model/ticket_event_meta.dart';
+import 'package:app_user/src/routing/app_coordinator.dart';
 import 'package:app_user/src/routing/app_router.dart';
 import 'package:app_user/src/routing/app_routes.dart';
 import 'package:flutter/material.dart';
@@ -8,18 +8,20 @@ import 'package:go_router/go_router.dart';
 import 'package:minglit_kit/minglit_kit.dart';
 
 final homeCoordinatorProvider = Provider<HomeCoordinator>((ref) {
-  return HomeCoordinator(ref.read(goRouterProvider));
+  return HomeCoordinator(
+    ref.read(goRouterProvider),
+    ref.read(appCoordinatorProvider),
+  );
 });
 
 class HomeCoordinator {
-  HomeCoordinator(this._router);
+  HomeCoordinator(this._router, this._app);
 
   final GoRouter _router;
+  final AppCoordinator _app;
 
   // Fix #634: goToLogin을 home_coordinator로 이동 — auth_coordinator 직접 참조 제거
-  void goToLogin({String? from}) {
-    _router.go(LoginRoute(from: from).location);
-  }
+  void goToLogin({String? from}) => _app.goToLogin(from: from);
 
   void pushNotificationCenter() {
     unawaited(_router.push(const NotificationCenterRoute().location));
@@ -43,9 +45,7 @@ class HomeCoordinator {
   }
 
   // Fix #404: Coordinator-based navigation for home route
-  void goToHome() {
-    _router.go('/');
-  }
+  void goToHome() => _app.goToHome();
 
   void pushPrivacy() {
     unawaited(_router.push(const PrivacyRoute().location));
@@ -53,18 +53,6 @@ class HomeCoordinator {
 
   void pushBlockedPartners() {
     unawaited(_router.push(const BlockedPartnersRoute().location));
-  }
-
-  // Fix #852: Navigate to ticket QR via coordinator
-  // — removes cross-feature import
-  // Fix #1526: eventMeta 추가 — 보딩패스 이벤트 정보 표시
-  void pushTicketQR(String ticketId, {TicketEventMeta? eventMeta}) {
-    unawaited(
-      _router.push(
-        TicketQRRoute(ticketId: ticketId).location,
-        extra: eventMeta,
-      ),
-    );
   }
 
   // Fix #1630: context.push() in AppBar 콜백은 내부 Navigator context를 사용해
@@ -79,9 +67,7 @@ class HomeCoordinator {
     unawaited(_router.push(const MyPageRoute().location));
   }
 
-  void pushEventDetail(String eventId) {
-    unawaited(_router.push(EventDetailRoute(eventId: eventId).location));
-  }
+  void pushEventDetail(String eventId) => _app.pushEventDetail(eventId);
 
   void pushPartnerEvents({
     required String partnerId,
